Update README.md
The data being analyzed for this project are measurements of human activity recorded by smartphones and stored at the following website:

http://archive.ics.uci.edu/ml/datasets/Human+Activity+Recognition+Using+Smartphones
 
The r script run_analysis.R reads the files from the directory UCI HAR Dataset.  It then merges the files together, extracts only the columns associated with the mean and standard deviations of the movements, appropriately labels the activities, and rewrites the file into a tidy data set.